requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.119.0
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httptools==0.7.1
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
streamlit==1.29.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
# uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
